Add files via upload
This is a simple Pay Scale model for employees based on the experience they have, and how the company offers salary as per the experience an candidate has. 

Now here we are to make model that could provide us more or less the exact same result we want. Here, I have simple linear regression technique on historical data of a company.

Thank you!!!